Cartbranch

+++ src/main/java/com/estore/ecomerce/controller/CategoryController.java
@@ -1,9 +1,12 @@
 package com.estore.ecomerce.controller;
 
+import com.estore.ecomerce.domain.Client;
 import com.estore.ecomerce.dto.CategoryResponse;
+import com.estore.ecomerce.security.ApplicationRole;
 import com.estore.ecomerce.service.CategoryService;
 import com.estore.ecomerce.service.FileUploadService;
-import com.estore.ecomerce.service.ImageService;
+import com.estore.ecomerce.service.abstraction.IUserService;
+
 import java.net.URISyntaxException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,6 +18,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.multipart.MultipartFile;
 
+import javassist.NotFoundException;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(path = "/api/v1/category")
@@ -24,16 +29,21 @@ public class CategoryController {
     @Autowired
     private CategoryService service;
     private final FileUploadService fileUploadService;
-    private final ImageService imageService;
+    private final IUserService userService; 
 
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<?> createCategory(
             @RequestPart(value = "image", required = false) MultipartFile image,
             @RequestPart(value = "category", required = true) CategoryResponse category)
-            throws URISyntaxException {
-           
-        ResponseEntity<?> response = service.addCategory(category, fileUploadService.uploadImageProfileToDB(image));
-        return new ResponseEntity<>(response.getBody(), response.getStatusCode());
+            throws URISyntaxException, NotFoundException {
+        Client client = (Client) userService.getInfoUser();    
+
+        if(client.getAuthorities().contains(ApplicationRole.ADMIN)){
+            ResponseEntity<?> response = service.addCategory(category, fileUploadService.uploadImageProfileToDB(image));
+            return new ResponseEntity<>(response.getBody(), response.getStatusCode());
+        }else{
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Request is FORBIDDEN. You havent the permission enough for the action");
+        }   
     }
 
     @GetMapping("")
@@ -61,17 +71,28 @@ public class CategoryController {
     public ResponseEntity<?> update(@PathVariable Long id, 
              @RequestPart(value="image",required=false) MultipartFile image,
              @RequestPart(value="category", required=true) CategoryResponse entity) 
-            throws URISyntaxException{
-      
-        ResponseEntity<?> response = service.update( id,entity,fileUploadService.uploadImageProfileToDB(image));
-        return new ResponseEntity<>(response.getBody(), response.getStatusCode());
+            throws URISyntaxException, NotFoundException{
+        Client client = (Client) userService.getInfoUser();    
+
+        if(client.getAuthorities().contains(ApplicationRole.ADMIN)){
+            ResponseEntity<?> response = service.update( id,entity,fileUploadService.uploadImageProfileToDB(image));
+            return new ResponseEntity<>(response.getBody(), response.getStatusCode());
+        }else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Request is FORBIDDEN. You havent the permission enough for the action");
+        }   
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) throws EntityNotFoundException {
-        service.delete(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    public ResponseEntity<?> delete(@PathVariable Long id) throws EntityNotFoundException, NotFoundException {
+        Client client = (Client) userService.getInfoUser();
 
+        if(client.getAuthorities().contains(ApplicationRole.ADMIN)){
+            service.delete(id);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Request is FORBIDDEN. You havent the permission enough for the action");
+        }
+        
     }
 
     @GetMapping("/active")
